+++ front/pages/api/w/[wId]/assistant/agent_configurations/[aId]/index.ts
@@ -2,16 +2,16 @@ import { isLeft } from "fp-ts/lib/Either";
 import * as reporter from "io-ts-reporters";
 import { NextApiRequest, NextApiResponse } from "next";
 
-import {
-  createOrUpgradeAgentConfiguration,
-  getAgentConfiguration,
-} from "@app/lib/api/assistant/configuration";
+import { getAgentConfiguration } from "@app/lib/api/assistant/configuration";
 import { Authenticator, getSession } from "@app/lib/auth";
 import { ReturnedAPIErrorType } from "@app/lib/error";
 import { apiError, withLogging } from "@app/logger/withlogging";
 import { AgentConfigurationType } from "@app/types/assistant/agent";
 
-import { PostOrPatchAgentConfigurationRequestBodySchema } from "..";
+import {
+  createOrUpgradeAgentConfiguration,
+  PostOrPatchAgentConfigurationRequestBodySchema,
+} from "..";
 
 export type GetAgentConfigurationResponseBody = {
   agentConfiguration: AgentConfigurationType;
@@ -82,18 +82,11 @@ async function handler(
         });
       }
 
-      const { name, pictureUrl, status, action, generation, description } =
-        bodyValidation.right.assistant;
-
-      const agentConfiguration = await createOrUpgradeAgentConfiguration(auth, {
-        name,
-        description,
-        pictureUrl,
-        status,
-        generation: generation,
-        action: action,
-        agentConfigurationId: req.query.aId as string,
-      });
+      const agentConfiguration = await createOrUpgradeAgentConfiguration(
+        auth,
+        bodyValidation.right,
+        req.query.aId as string
+      );
 
       return res.status(200).json({
         agentConfiguration: agentConfiguration,
